Removing EF warnings. Converting to INT

+++ src/FairPlayCombinedSln/FairPlayCombined.DataAccess/Models/FairPlayShop/Product.cs
@@ -32,7 +32,7 @@ public partial class Product
 
     public int QuantityInStock { get; set; }
 
-    public byte ProductStatusId { get; set; }
+    public int ProductStatusId { get; set; }
 
     [Required]
     public string OwnerId { get; set; }
